chore: rename version migrator file to correct version number
